extract-linkedin-premium-users-from-linkedin-groups/agentic/crewai/pyproject.toml
[project]
name = "linkedin-premium-extractor-crewai"
version = "1.0.0"
description = "Extract LinkedIn premium members using CrewAI and ConnectSafely.ai"
authors = [
    { name = "ConnectSafely.ai", email = "support@connectsafely.ai" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "crewai[google-genai]>=0.80.0",
    "crewai-tools>=0.12.0",
    "streamlit>=1.39.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.154.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
